Add EGLContext validation to fix bug ANDROID-1679
Segmentation fault will occur when using an invalid EGLContext,
such as getting an EGLContext by "(EGLContext)3".

This validation process first uses msync system call
to check for unmapped memory pointer.

Moreover, Because existing properties(like dpy, config, version, etc.)
are hard to identify if this is an egl_context_t instance,
a magic property is introduced into the class egl_context_t,
which helps to see if it does represent an egl_context_t instance
for most basic accidental cases,
though not feasible to test for deliberate forgeries.

Once the pointer is confirmed valid,
invoke the isValid method of egl_context_t,
which compares magic value to the constant MAGIC to do validation.

Change-Id: If1b884134af57f703b51eb483c4bd0a85c78ec5a
Original-Change-Id: Iebebf4238498c667319c62469baa5d8538127b2d
Signed-off-by: Yi <yi.xie@intel.com>
Signed-off-by: Oliver McFadden <oliver.mcfadden@intel.com>
Signed-off-by: Russell Webb <russell.webb@intel.com>